add googletest submodule and refactor examples directory
