+++ .gitmodules
@@ -1,3 +1,6 @@
 [submodule "benchmark/vendors/benchmark"]
 	path = benchmark/vendors/benchmark
 	url = https://github.com/google/benchmark.git
+[submodule "test/vendors/googletest"]
+	path = test/vendors/googletest
+	url = git@github.com:google/googletest.git
