+++ example/omp_policy.cpp
